Separate notion of Context and ContextBuilder
