+++ src/Docker/Tests/Context/context-test/Dockerfile
@@ -1,0 +1,3 @@
+FROM base
+
+RUN echo 'hello world'
